A mostly-simple State class.
The idea is that there are both primary and secondary states.  The
primary state is what's used to determine which handler deals with
the state (are you logging in?  are you chatting? etc.) and the
secondary state is used by that handler for whatever internal
knowledge it needs.  The secondary state is wiped every time you
make a primary state transition.