Single-instantiate the die roller/configurator.
There's no reason to keep creating and destroying these objects.
Instead, the server has a single instance of each one, which are
then reused.  This reduces CPU usage (not that Giles is exactly
a processing beast) at the possible expense of a bit of memory
to keep objects around that aren't being used.  I'll live.