Start on primary functionality: game_master/RPS.
Right now the game master simply manages sessions and routes commands
to the instances of games under its control.  You can also create a
new rock-paper-scissors instance, but it will do nothing but report
for duty if you send it a command.  Still, this is a major step to
actually supporting real games.

The design of this went back and forth a bit, and I had a complicated
Session object that sat in between instantiations, but... the whole
point of objects is that they ARE instantiations, so if we make a
RPS() object it should be an instance of the game.  Silly Phil,
overcomplicating things.