Support repeated games in the same channel.
After some discussion with DG and pineapple, I came to the conclusion
that allowing repeat games in a channel is okay.  The stipulation is
that the channel was itself initially created by a game; presumably
if you're in there chatting about a game, you're fine with playing
another one.  People can't gamebomb other folks' channels, and global
will never have RPS playing on it.

I also fixed a bug where channels created by games were persistent
even after the game ended, which is a bad thing, as you'd never be
able to create a new game with that name because of it.