Major nomenclature change: session -> table.
I found myself using them almost interchangeably as I worked on the
generic implementation, and I think it's a worthwhile change and
abstraction.  Of course, on the server tables don't really exist
in a /location/--you can access them from anywhere--but from the
/game's/ perspective it's useful.  Anyway, it's basically a NOOP
other than a whole bunch of variable name changes, some text string
ones, and the fact that the "session" command is now "table" (but
everyone should be using / anyway).