Add cleanup and notifications of new games.
Before, if you were in a channel and a new game was started with
that channel, you didn't see it unless you happened to be in scope
of that game's creation.  Now it sends a notification to the channel.

There's also a cleanup function in the game_master now; the server
will crash at the moment when someone disconnects, because RPS does
not implement that cleanup function.  The about-to-land generic
game structure /does/ use that cleanup function, however.