Major functionality change: generic Game.
This commit marks the addition of a generic Game class.  You can't
actually /play/ it, but you can join it (if it's been set to the
right mode), leave it, list the players in it, and so on.  I will
now convert RPS to using the class, which should get us back to
having a working game... albeit one that's a lot more featureful.