Add show() implementation for RPS.
This was more for fun than because of any real need, but it /does/
tell you which players have picked their throws, which haven't,
and whether seats are missing, so I suppose it could be used as a
quick 'status check'.  Of course, who will be playing RPS outside
of a testing environment anyway?