Y: Default to need_players.
You can still setup the game; just send a /table setup and it'll
switch modes.

This is a better choice, because it lets players get started playing
immediately, but the configuration can be tweaked if they want too.
This plays into my long-term plans for doing things like allowing
major alterations to the rules of (for example) Lost Cities, but
making it easy to play the default game quickly.