Major new feature: public/private games, "active".
This is separate from their scope, which is just where they are
announced.  A public game shows up in the new 'game active'
subcommand; a private game only shows up if you're on its channel
or are privileged enough to see all games.  (Right now, no one
is.)

Yes, that means you can now actually get a list of all active
games on the server, which is quite useful.