server, player: Generate timestamps in the server.
Once a second we update the server timestamp.  This is overkill, as
it's only precise to the minute, but that means we get ~1s accuracy
as to the minute transition.  Also get rid of all of the calls to
strftime() in player.py, instead using this cached copy.

Yes, this is probably a micro-optimization, but it makes me happy.