Set: Order the layout in columns rather than rows.
This may seem silly, but what it does is keep the cards from jumping
around when new ones are added due to no set being found.  When it
was ordered by rows, that shifted all cards on the second row to
the left one and on the third row to the left twice.  If you were
in the middle of typing out a set, the values likely changed.  Now
they are stable when cards are added, which is a clear win.