Set: Move to using bitfields; structures.
This should have been done in several steps, but I ended up conflating
them.  Anyhow, we now use bitfields to determine the third card of a
set when given any two other cards, and that functionality is also used
by is_a_set() (which is a single line now) and by the function that
determines whether or not there are any more playable sets.

The conflation is that I /also/ switched to using structures to hold
all of this information rather than the previous strings where I did
stupid pet tricks depending on what the string was.  Both worthwhile
changes, but the result is that two changes that could have been
separated are not.  Ah, well.