Major new feature: Capture Go.
Capture Go makes use of a new Goban class that understands how to
do Go captures (and will be used for Gonnect and, eventually, an
actual Go implementation).  Capture Go does not allow passing, and
defaults to single-capture.  There's still a thing or two that
could be done with the goban to make it more efficient (see the
TODO), and it doesn't understand ko/superko, but it's a totally
playable implementation.