Major new feature: Gonnect (mostly).
This is a mostly-full Gonnect implementation.  It supports odd-shaped
boards, just like Capture Go, for the fun of it; it also supports
directional goals (Horizontal vs. Vertical), but the board must be
square if you have directional goals.  (This makes me think that
I should, in fact, support non-square boards as a handicap, both
here and in Hex... something to think about later.)

What it does /not/ support is resolving ko, as the Goban class does
not yet handle resolving ko.  This will have to be added later to
make it fully worthwhile playing (as you can get into a "ko fight"
and stall the game out as-is).