BT/CG/Cr/Go/Hex/Y: Use player_name more.
All of these games have a find_winner() function that returns the
name of the winner.  They should return the player's name, not the
player object, in case the player has left.  (It turns out that
for all of these games this is an irrelevant distinction, as
winning can only occur right after a player makes a move, and only
the moving player can possibly win... but if someone copies one of
these games to use for their own that does not have this distinction,
they could encounter a very odd bug.  We don't want to cause bugs
we can avoid.)