Major new feature: Satisfy the AGPL, config.
There is now a mandatory configuration files, giles.conf.  The only
option you absolutely MUST set is source_url, which should be a link
to a copy/repository/whatever of your custom Giles version.  This
satisfies the requirements of the AGPL.  You can also configure the
name of the server, which is nice, but you don't have to.

.gitignore is updated to ignore giles.conf for obvious reasons.