New library/function: trick.py, handle_trick().
handle_trick() is meant to handle 98% of your trick-taking game needs.
Given a list (or Hand) it can dig into, with each element having a
.suit variable and comparability, it will determine the winning card.
It can handle trumps (pass in the trump suit) and, as an added bonus,
can handle either "first wins" or "last wins" with duplicate-valued
cards.

Not tested yet, mind, although it /seems/ correct... hopefully I'll
get a chance to work on Whist tonight before the hurricane hits.