game_master: Seven games per line.
Now that the server has ten games, the line listing the games is
starting to get dangerously long.  For now we just do a newline
after every 7.  An odd number means that each line is "checkerboard"
in its coloration.

This should all be eventually replaced with more meaningful output,
but it'll do for now.