Major new feature: Whist.
That's right, Giles has its first card game.  Woo!  Whist was the
"easiest" to implement, in the sense that its scoring is without a
doubt the simplest of all trick-taking games I know, particularly
if you ignore honours (as you should, because honours are boring).

I've only tested this against myself as all four players, so I'm
sure there are subtle lingering bugs, but I have to say that it
works surprisingly well.