Whist/playing_card: Functionalize hand-to-string.
This is something that just about every playing-card game will want
to do anyhow, so why not make it a function?  I thought about putting
it in trick.py first, but right now trick.py is actually deck-agnostic;
you could potentially use it with a Tarot deck and have it still work.
This stringification, however, colorizes hearts and diamonds, and as
such is quite particular to French-suited cards.

Anyhow, this results in no functional changes, but allows me to make
use of the code again in later games.