Major new feature: Layouts.  Convert Whist.
This is something that's overdue: the concept of a layout, which is
simply a representation of the game space.  Different types of games
can use different layouts, but some are quite common, such as the
one implemented here, for four-player card games with North, East,
South, and West as players.  Note that the layout doesn't understand
anything about partnerships or the like; it just knows how to draw a
table and the cards on the table.

The real gain here is that I don't have to rewrite or cut-and-paste
all that guff for future 4p card games, but while I was at it I did
in fact make an enhancement: the last card played on the table is now
bolded.