Major new feature: 3p Hokm.
This makes the code quite a bit uglier, as 3p Hokm is not partnership,
which changes quite a bit of the internals... but it's actually less
ugly than I was expecting.

3p has been (relatively) tested, and supports both short and long
decks; the default is short, because that's the superior game.  4p
mode still starts, but a full game hasn't been played yet after the
changes.