New feature: Admin privileges.
There is a single admin password that any user can enter which is
kept in giles.conf.  Right now the only admin commands are 'on' and
'off,' which enable and disable admin privileges... but there's
nothing to do with those privileges just yet.

Since this new module was added after I started "crash-resisting"
the server, all calls to the admin manager are wrapped with crash
protection as well.