Major new feature: Load new games on the fly.
This was the real target for all of the admin work today.  Now you
can add a whole new game to a Giles server on-the-fly, and thanks
to the crash protection wrapped around the game-related commands
now, working on a running server should actually be no problem at
all.

Eventually we should support crazy things like actually reloading
the admin manager itself, and potentially the other managers as
well, but for now this covers 95% of the "heavy development" usage;
I no longer have to keep taking the test servers up and down while
I add new game functionality.  This is, in a word, super-awesome.
(Okay, maybe two words.)