admin_manager: new reload_by_name() feature.
Reloading game modules is good, but if one of their dependencies
changes (such as the new functionality I added to SquareGridLayout)
the reload doesn't reload /those/ dependencies.  Now you can, by
giving the name of the module via admin reload_by_name.  This is
not meant to be used to reload games or other modules, as it won't
replace the running instances, but is useful for fixes in utility
libraries.