Major new feature: Square Oust.
Another Mark Steere game.  This one has some pretty complicated bits,
having to deal with the sizes of adjacent opponent groups, but it seems
to work correctly.  One of the advantages of Giles play is that you can
do so on quirkily-shaped boards, as long as both dimensions are odd.