Objectify chat.
This provides only one immediate benefit: if something goes horribly
awry with the chat module, it's now caught as an exception and the
server doesn't crash.  This is a good thing!

Its real purpose, though, is to let chat and its objects get hot
reloading via the admin module, which is super-useful for adding new
functionality while the server stays up.