New feature: focus/unfocus.
This (currently undocumented, because documentation in Giles is a
complete and totel mess) feature allows you to focus on a specific
table.  Once focused, every command goes to that table unless you
preface it with a '/', in which case it gets parsed by the standard
parser. 'unfocus' (obviously) lets you unfocus... but of course it
needs to be typed as '/unfocus' for reasons which are hopefully
obvious.

This necessitated a couple of helper functions that probably should
have been written before, as there are now more things that care
about table state than before.  I should've made those separate
commits, but...