Break Game into Game and SeatedGame.
For some planned games that I want to write, in the vein of LORD and
MUDs, the idea of seats simply doesn't make sense.  As such, having
to subclass the previous Game class and then studiously keep people
from running all of the commands that have to do with seats seemed...
dubious.  So instead I've kept the stuff that's still essential for
all games (mainly scutwork, like making channels and looking at who
is playing) in game.py and made a new seated_game.py which has all
of the things that have to do particularly with traditional board
games: finding a seat, etc.

Functionally this provides no change whatsoever to how any existing
game works (modulo bugs), but it provides a platform for further
work.