Actually track the config filename.
game_master.py was blindly reloading giles.conf because, well, this
wasn't an option until recently.  This was seriously messing up a
major config change I'm in the middle of implementing.  This is still
rather hokey; the right thing now is probably to /just/ pass the
filename and do the right loading thing in server.py or somewhere else,
but it'll do for now.