Poison: Autoquaffing and handling results.
This is actually most but not all of the remaining logic.  I still need
to handle the 'pick a seat to quaff' phase, the 'pick a potion to toss'
phase for when you poison yourself, and the 'pick a seat to be first'
when you kill yourself, but you can actually win a (trivial) game now.