Support more or less piece rows in Breakthrough.
Reading a PDF discussing the game and strategies made me realize that there is
nothing sacred about the "two rows of pieces" for Breakthrough.  In fact, it
referred to a 10x10 board with 30 pieces per side, which would be three rows.
This modifies Breakthrough to support such configurations, although it still
defaults to 8x8 with two piece rows.  A new command "rows" (shorthand "ro") is
part of the Breakthrough configuration to support this change.

While I was at it, I also reduced some minimums; it turns out that 2x2
Breakthrough is still valid (although a trivial win for the first player), so
if someone wants to play that, why not let them?