Forty-One: Add bid decay.
We played a lot of 41 in my last job, and we started playing with a 'decay'
variant where, if the minimum bid number was unmet, the next hand would have
a lower minimum bid.  Once a bid was finally met, the value ratcheted back
to the original number.  This implements that variant.