A better comment about tick rate.
I found it deeply suspicious that ticks were *still* 20/s on a machine
that is much faster than the first one I wrote Giles on, so I did some
profiling and saw that, yup, all the time is spent in select.select()
based on the timeout parameter.  That makes a lot of sense, obviously.
But the extant comment implied that the rate might be variable based
on the machine you run this on.  Unless it's on a very underpowered
embedded processor, my guess is that select.select() will *always* be
the dominant time factor for Giles.  So make that clear for my future
self.