game: Add show_config.
There's currently no canonical way to look at the configuration of
games, and generally 'show' isn't the right place for it.  So this
adds that; each game can define a config_params that gives helpful
descriptions of each configuration option and the name of the
appropriate attributes.  This requires some Pythonic introspection,
but nothing too fancy.